Fix issue + separate for single responsibility
