+++ packages/core/src/services/integrations/pipedream/triggers.test.ts
@@ -9,7 +9,7 @@ import { updatePipedreamTrigger } from './triggers'
 import { Result } from '../../../lib/Result'
 import { IntegrationTriggerConfiguration } from '@latitude-data/constants/documentTriggers'
 import * as appsModule from './apps'
-import * as componentsModule from './components'
+import * as componentsModule from './components/fillConfiguredProps'
 import * as triggersModule from './triggers'
 import { BadRequestError, NotFoundError } from '@latitude-data/constants/errors'
 import * as factories from '../../../tests/factories'
